feat: add cards layout, change filters styles
