+++ src/enums/CardsLayout.ts
@@ -1,0 +1,4 @@
+export enum CardsLayout {
+  First = 'first',
+  Second = 'second',
+}
