+++ src/pages/StorePage/StoreCard/StoreCard.module.scss
@@ -1,5 +1,5 @@
 .imageWrapper {
-  max-width: 300px;
+  max-width: 80%;
   width: 100%;
   & > div {
     height: 0;
